Fixed a bug where an app installed from zip file is deleted by foxx manager.

+++ js/client/modules/org/arangodb/foxx/manager.js
@@ -264,8 +264,10 @@
       throwFileNotFound("Cannot find file: " + appInfo + ".");
     }
     var filePath;
+    var shouldDelete = false;
     if (fs.isDirectory(appInfo)) {
       filePath = utils.zipDirectory(appInfo);
+      shouldDelete = true;
     }
     if (fs.isFile(appInfo)) {
       filePath = appInfo;
@@ -274,11 +276,13 @@
       throwBadParameter("Invalid file: " + appInfo + ". Has to be a direcotry or zip archive");
     }
     var response = arango.SEND_FILE("/_api/upload", filePath);
-    try {
-      fs.remove(filePath);
-    }
-    catch (err2) {
-      arangodb.printf("Cannot remove temporary file '%s'\n", filePath);
+    if (shouldDelete) {
+      try {
+        fs.remove(filePath);
+      }
+      catch (err2) {
+        arangodb.printf("Cannot remove temporary file '%s'\n", filePath);
+      }
     }
     if (! response.filename) {
       throw new ArangoError({
